requirements.txt
fastapi==0.115.0
uvicorn==0.38.0
sqlalchemy==2.0.45
pydantic==2.12.5
python-dotenv==1.2.1
httpx==0.28.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
